factor out different flavors of lims queries
