+++ allensdk/internal/api/queries/mtrain_queries.py
@@ -1,7 +1,7 @@
 from typing import Optional, List
 import pandas as pd
 from allensdk.internal.api import PostgresQueryMixin
-from allensdk.internal.api.queries.lims_queries import (
+from allensdk.internal.api.queries.utils import (
     build_in_list_selector_query)
 import logging
